Amazon-Delivery-ML/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.1
numpy==2.2.6
scikit-learn==1.7.2
joblib==1.5.1
python-multipart==0.0.20
PyJWT==2.10.1
